requirements.txt
cachelib==0.7.0
click==8.1.3
Flask==2.1.2
Flask-Session==0.4.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
mysqlclient==2.1.0
psycopg2==2.9.3
python-dotenv==0.20.0
regex==2022.6.2
SQLAlchemy==1.4.36
Werkzeug==2.1.2
google-cloud-secret-manager

